Fix slow character switch for empty craft Level
Regular expression checking turned out to be a bad idea for sliders.
Sliders have to convert binding values to doubles and this works poorly
if the source is not easily convertible, e.g. null or "". This would
eventually bring the slider to a crawl. Now, however, it no longer
validates.