Add completion overview
Completion overview should open a new window with a table showing the
completion state of all areas for all characters. Characters will be on
the horizontal axis, areas on the vertical axis. Characters will ideally
use the ordering from the character list. Areas will ideally use the
ordering from the reference list, but to start with will simply be
alphabetical.