Convert to Git
This is an empty commit denoting the point in time this repository was
converted from Mercurial to Git.

The conversion was carried out with reposurgeon 3.20 [1].

* Commit messages have been cleaned up to comply with the standard Git
  commit message guidelines, in particular the 50/72 rule.

* Some messages have been clarified.

* Some commits have been annotated with editorial notes wrapped in [[ ]].

* Empty commits have been removed.

* Some commits have been squashed.

The reason for the conversion is my personal preference for Git and the
ability to eliminate Mercurial as a dependency.

The conversion, both as means to an end and an experiment in itself, has
been mostly succesful, with a few hitches. One example of an issue is
the dos2unix-ification of all files, which unexpectedly ruined images;
these had to be manually edited back in at the correct point in history.
Another is a commit switching from master, where it belonged, to a
branch that was rooted in said commit; at that time, I had spent enough
time on the conversion and no longer cared enough to resolve that.

[1] http://www.catb.org/esr/reposurgeon/