Update WPFToolkit to 2.4
Numerous changes have been made to WPFToolkit between 1.8 and 2.4, but
the major change seems to be the introduction of the Avalon control.
This may or may not become relevant in the future, but for now only the
many bugfixes made to the number controls are important. Also,
SelectAllOnGotFocus now defaults to True.

The toolkit is now split into several separate DLLs. Only the core is
included here -- the rest can be added if they become necessary.

The way the DLL embedding happens has been changed slightly. 1.8 was
treated as a resouce of type File, but after spending several hours
trying to remember how it was referenced in VS, now the DLL is simply
added straight to the solution via drag-and-drop, with its build type
set to embedded. This should be easier to return to in the future.
Since the DLL path is now different, the assembly resolution has had to
be updated.

Some implementation notes:

* The resource stream name is
  "{namespace}/relative/path".replace("/", ".") [1].
* Return null for the {namespace}.resources assembly [2].
* Invalid resource stream names will no longer be attempted loaded.

[1] http://stackoverflow.com/a/11542163/482758
[1] http://stackoverflow.com/a/10600027/482758