Fix completion overview
Closes #35.

At some point during the lifetime of the Mercurial repository, the
completion overview ceased to render character columns. No cause for
this could be discerned; in fact, everything suggests that the original
implementation never worked, although history is evidence to the
contrary.

The problem has been narrowed down to a timing of the call to
Window.Show(): if called before *all* columns have been added to the
data table, the table does not update when new columns or rows are
added. Calling Show() before adding rows is not problematic. This is
certainly related to the runtime generation of columns, somehow.

This patch moves the call to Show() to right after all columns have been
generated and before rows are generated to preserve UI responsiveness.