Automate distribution creation
This patch adds a script, adapted from a batch file [1], that builds a
release binary and prepares it for distribution. The script is a
make-shift makefile, because Make is an extra dependency in MSYS and the
build is fast enough that incremental building is not important.

It's written in Bash because the Batch language is opaque and makes path
resolution harder than necessary. However, only Windows is supported;
the script must be run from an MSYS-like environment.

[1] http://stackoverflow.com/a/5669820/482758