Clean up ignore file
The entries in this file were mostly the shotgun-ignores found in places
like the ignore-file repository on GitHub [], although the actual source
is unknown. It is better to ignore files on a case-by-case basis.

There were also a few entries from a completely unrelated project that
should never have been there.

[1] https://github.com/github/gitignore