Manually track non-core story chapter completion
Closes #46

Add a rudimentary completion tracker for non-core story arcs at chapter
granularity, so it can be used to track exact progress and BLC key
rewards.

A lot of neat nice-to-have UX is not included. Most importantly,
chapters can be marked completed out-of-order and there is no episode or
storyline summary. It seems like a decent value-add for now, though.