Fit window height to content
ref #46

Rather than offload the overhead of resizing the window to fit the story
tab contents to the user while avoiding pre-allocating a lot of empty
space in other tabs, automatically adjust the height to the content.
Restrict the minimum height to fit the smallest view and the maximum
height to fit the largest view so the window doesn't expand indefinitely.