Support fixed character listing order
Assign each character a "sort order" in the range [1;99] with default 1.
Expose this value in the character list and choose it as the default
sort description. Leave the grid freely sortable by users so they can
temporarily override the default sort order as needed.

The sort order is just a dumb weight mechanism: greater values weigh
more and "sink" further towards the bottom. It can only be manipulated
by selecting a character and manually setting the sort order, which is
somewhat cumbersome but is very simple and reliable, whereas
drag-and-drop supposedly is "difficult" with (the default) DataGrid. I
think this inconvenience is OK: I expect permanent rebalancing to need
to happen quite rarely.

ref #40