Remove Extended WPF Toolkit
The Extended WPF Toolkit now appears to be the only thing that stands in
the way of moving to .NET 5. The library has a for-pay .NET Core 3
version but no mention of .NET 5, and besides, the IntegerUpDown control
is the only one I ever used. Reimplement the IntegerUpDown control from
scratch and remove the library; thereby also halving the binary size.