Move JumpList out of ViewModelMain
There is no JumpList registered for the test runner assembly so
initializing the JumpList from a test runner fails in all sorts of ways,
making it impossible to test file loading.

It's possible to make the implementation robust against these errors but
since they're artificial it's rather pointless. Instead, just move the
JumpList implementation into the main App XAML file, where it makes a
whole lot more sense than inside a VM anyway.

This creates a conflict between ViewModelMain's file path constructor
for opening a file, the outcome of that operation, and the JumpList.
That constructor seems awkward, too, however, so move the argument to
the Open method, raise that method's visibility, and make it return a
status we can base the JumpList behaviour on.