Change completion overview data context to code-behind
Make the code-behind be the data context instead of the data table so we
can bind to named properties on a model in our control, so we can bind
to additional properties. Half-way towards MVVM but without the somewhat
superfluous dedicated VM. This is preparation for introducing a row
style selector.

Also Remove a redundant nesting in the completion overview view.