Fix some completion objectives not changing completion state
When adjusting progress on heart and skill completion objectives, the
area's completion state doesn't change. Fix this.

I have a nagging feeling that I made this specifically ignore hearts and
skills due to some areas' completion objectives but I have no supporting
evidence and can find no indication that this would be necessary. I've
tested with cities, which are the only areas with neither hearts nor
skills, and the behaviour is as desired.