Use new Wiki profession icons
Charrmander has always used the Wiki's "tango" icons because, well, they
were the only ones available. The tango icons follow a particular
aesthetic, mildly stylized, though whether this was ever intentional or
merely a historical curiosity is unknown.

The tango icons are handcrafted by community members. This has the
advantages that there is no legal question about distributing in-game
art and that the icons can apply more detail. It has the disadvantage
that crafting these icons is time consuming and error prone.

In the month leading up to the End of Dragons release, the Wiki replaced
every 20px version of the tango profession icons with alternatives
extracted from the game data, scaled down, and colourized. That is, the
new icons are derivatives. Besides the legal grey area that comes with
that, the aesthetic of the official game art is generally reminiscent of
the Downs syndrome Black Lion lion so all the new profession icons look
like wine gums.

But besides all the reasons to dislike the new profession icons, the
tango icons have a host of their own issues. Especially the End of
Dragons elite specialization icons are of pitiful quality, with too
light hues and details that vanish at low resolution. Some old icons
have pretty jarring issues, too, like the engineer bomb's fuse being
completely misplaced or the guardian shield being stretched
horizontally.

Commit 05d1dfb (Add End of Dragons specializations, 2022-01-29) added
the new specializations using the new, low-grade tango icons to preserve
Charrmander's internal aesthetic. This commit replaces the tango
profession icons with the game art wine gums, whose aesthetic is clear
and consistent if unmatched with Charrmander. These icons are available
in the new "Profession icon" template [1].

[1] https://wiki.guildwars2.com/wiki/Template:Profession_icon